reactions api integration, chore: refactor and collate verification/comments/reactions controller methods
