+++ desci-server/src/controllers/attestations/index.ts
@@ -1,12 +1,6 @@
 export * from './show.js';
 export * from './claims.js';
+export * from './comments.js';
 export * from './reactions.js';
-export * from './addComment.js';
-export * from './getComments.js';
-export * from './addReaction.js';
-// export * from './addVerification.js';
 export * from './recommendations.js';
-export * from './removeComment.js';
-export * from './removeReaction.js';
-// export * from './removeVerification.js';
 export * from './verification.js';
